+++ plugins/TooltipNotify/src/Utils.h
@@ -5,4 +5,4 @@
 #pragma once
 
 BOOL IsNt50();
-void TruncateWithDots(TCHAR* szString, int iNewLen);
+void TruncateWithDots(TCHAR* szString, size_t iNewLen);
